refactor: split out app into multiple router modules
